requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
elasticsearch==7.13.4
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.0.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
